Reorganize code structure for better readability and delete react logo file
